requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.34.77
botocore==1.34.162
celery==5.3.6
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.20
django-environ==0.11.2
django-filter==23.5
django-guardian==2.4.0
django-storages==1.14.2
djangorestframework==3.14.0
et_xmlfile==2.0.0
Faker==37.1.0
jmespath==1.0.1
kombu==5.5.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
redis==5.0.1
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
